Implement speed hack for fast insertion of large amounts of text via the InsertTextCommand.
Conversion from display index to string index can be slow, so only calculate the string index from the display index for the first character of the text. For the consecutive characters the stringIndex is calculated by incrementing the stringIndex of the previous character. When the previous character was an EOL character it is set back to 0. This avoids the expensive call to Line.DisplayIndexToStringIndex(...)
This reduces the first time population duration of the FastColoredTextBox for large files from 57 seconds down to 2 seconds.